Updated the signature of DelphiRule.visit
The only place where "visit" method is called is a few lines below in
"visitAll" method. It passes DelphiPMDNode and RuleContext into it. Also
it does not expect any return value. So I updated the signature of
"visit" to match the way it is used. I see no bonuses in passing of
RuleContext as a plain Object and adding of line "return data;" in every
overloaded method.