Fixed xml parsing errors that were causing issues on loading the server.
The rules xml files were missing a couple of closing tags, likely
because of a merge conflict that was incorrectly resolved.

Note for future: If the server logs say that 'DelphiRule.rule' does not
exist, that probably just means the .xml files for rule definitions is
invalid, it's a bit of a misleading error.