Rule working well and is configurable, fixed some null pointer errors.
Made a few minor changes to the code logic, as the array way going out
of bounds due to .size returning the size of the list from 1 not 0 so
those have been fixed.

Some files may also not have any children under the implementation
declaration, so added a check for this to remove null exceptions,
otherwise looking good!