Fix performance regression triggered by SonarQube 9.8
In 9.8, the `is` file predicate now calls `Files.isSameFile` instead of
doing a basic path check.
Not desirable for us when it's actually hitting the disk (!) thousands
of times per call to `DelphiProjectHelper.getFile`. The information we
care about is already in memory though, and all we needed was the
path check, so we've just swapped over to the `hasPath` predicate and
effectively returned to the old behavior.