Rework interface for creating Delphi test files
The existing `UnsupportedOperationException`s thrown in the
`DelphiTestFileBuilder` class indicated that the class hierarchy design
was not accurately reflecting the requirements. Furthermore, the
`CheckVerifier` interface was accepting `DelphiTestFileBuilder<?>` in
its `onFile` method, which was both clunky and unneccesarily specific.

This commit refactors it all into a new hierarchy.

```
DelphiTestFile
      |
      v
AbstractDelphiTestFile -------------------
      |                                  |
      v                                  v
AbstractDelphiTestFileBuilder    DelphiTestResource
      |                   |
      v                   |
DelphiTestProgramBuilder  |
                          |
                          |
DelphiTestUnitBuilder <---|
```

Co-authored-by: Joshua Gardner <jgardner@integradev.com.au>