Skip integration tests by default
Because integration tests spin up a SonarQube server, they share
SonarQube's Java requirements.

SonarQube only runs under Java 17 (no other versions are supported),
meaning that the integration tests require the developer to be running
JDK 17 *exactly*.

This is kind of a weird and inconvenient limitation to impose on
contributors, so we're skipping the integration tests by default.