Handle aliases when casting to `*TypeImpl` classes
This is a little bit of a "gotcha".

Now that type aliases are transparent to the aliased type, we need to do
a little extra work before internally casting a type to `Impl` for
mutable operations.

Otherwise, you might try to cast a generated alias implementation to
some unrelated type.