Make imaginary tokens unmatchable
This is a silly mistake that we inherited in the Delphi grammar.

We've been defining our imaginary tokens with actual values, and not
thinking much of it. However, this technically means that they could be
matched in source code by the lexer and wreak all sorts of havoc.

Declaring imaginary tokens in a `tokens` block is a clean and easy way
of preventing them from ever being matched.