Improve support for program files omitting the `program` header
This was partially implemented in #159, but stopped at the parsing
level.

Throughout the analyzer, there were unsafe assumptions about the
presence of a file header. The major area that needed extra
consideration was symbol table construction.

It turns out that the default name of a Delphi program is `PROGRAM`, so
this name ends up being used if the header is omitted. See
`DefaultProgramName.dpr` for an example of a valid Delphi program that
relies on this behavior.