Make the grammar stricter around `string` and `file` types
Previously, any place where a `typeReference` (aka `TypeReferenceNode`)
could appear, we would also allow `string` and `file`.

This is unnecessarily permissive, and in invalid code we could see
`StringTypeNode` and `FileTypeNode` in unexpected places where we're
expecting (and casting to) `TypeReferenceNode`, leading to gnarly errors
later in the scan.

The grammar has been tightened up a bit now, and we only allow
`stringType` and `fileType` in places where they can legally appear.