Persist routine directives into `ProceduralType`
This was prompted by the realization that `varargs` weren't consistently
persisted into procedural type data, causing non-routine procedural
types to never be recognized as variadic.

Now we simply persist all of the directives, which will also make it
easier to factor calling conventions when comparing procedural types
for overload resolution.