Prevent double-resolution of property specifiers
Property specifiers `read`/`write`/`implements`/`stored` were being
resolved twice, which was unnecessary because they should be resolved as
part of the top-level resolution of the `PropertyNode`.

In a funny twist, it turns out we were actually relying on that second
visit to resolve non-invocable expressions in `read`/`write`/`stored`
specifiers. This is now handled properly in the top-level resolve.