Disambiguate index property read/write methods correctly
When a property has an `index` specifier, its `read` specifier must be
a function with an `Integer` final argument, and its `write` specifier
must be a procedure with an `Integer` second-last argument. We were
not previously modelling this behaviour, causing `read` and `write`
specifier methods to fail to resolve when an `index` property is present.